+++ docs/release_notes.rst
@@ -4,7 +4,7 @@ Release Notes
 v2.1.0
 ------
 
-* Add ``stor.S3Path.restore(days, tier)`` to restore objects from Glacier to S3.
+* Add ``stor.S3Path.restore(days, tier)`` to restore *single* objects from Glacier to S3.
   Paired with this change are two specific exceptions related to S3 restores,
   ``RestoreAlreadyInProgressError`` (raised when restore has already started)
   and ``AlreadyRestoredError`` (raised when restoring object already in S3).